+++ Source/General/GlobalCommands.vb
@@ -334,9 +334,9 @@ Public Class GlobalCommands
                 form.Doc.Write("Active Contributors", "Patman86, Valdiralita")
                 form.Doc.Write("Retired Authors", "stax76, JKyle, 44vince44, DJATOM, Revan654, NikosD, jernst, Brother John, Freepik, ilko-k, nulledone, vanontom")
                 form.Doc.Writer.WriteRaw("<hr>")
-                form.Doc.Write("Links", "<a href=""https://github.com/staxrip/staxrip"">Source on GitHub</a>" &
-                                         "<br><a href=""https://github.com/staxrip/staxrip/blob/master/Changelog.md"">Changelog on GitHub</a>" &
-                                         "<br><a href=""https://github.com/staxrip/staxrip/wiki"">Wiki on GitHub</a>" &
+                form.Doc.Write("Links", "<a href=""https://github.com/m00nxx/StaxRip2"">Source on GitHub</a>" &
+                                         "<br><a href=""https://github.com/m00nxx/StaxRip2/blob/master/CHANGELOG.md"">Changelog on GitHub</a>" &
+                                         "<br><a href=""https://github.com/m00nxx/StaxRip2/wiki"">Wiki on GitHub</a>" &
                                          "<br><a href=""https://discord.gg/uz8pVR79Bd"">StaxRip Community on Discord</a>")
                 form.Doc.Writer.WriteRaw("<hr>")