+++ Source/My Project/AssemblyInfo.vb
@@ -1,12 +1,12 @@
 ﻿Imports System.Reflection
 Imports System.Runtime.InteropServices
 
-<Assembly: AssemblyTitle("StaxRip")>
-<Assembly: AssemblyCompany("StaxRip Authors")>
-<Assembly: AssemblyProduct("StaxRip")>
-<Assembly: AssemblyCopyright("Copyright (C) 2002-2026 StaxRip Authors")>
+<Assembly: AssemblyTitle("StaxRip2")>
+<Assembly: AssemblyCompany("StaxRip2 Contributors and StaxRip Authors")>
+<Assembly: AssemblyProduct("StaxRip2")>
+<Assembly: AssemblyCopyright("Copyright (C) 2002-2026 StaxRip Authors; StaxRip2 modifications Copyright (C) 2026 StaxRip2 Contributors")>
 <Assembly: ComVisible(False)> 
 <Assembly: Guid("54d5dc0f-6ba7-42f0-a82f-9c71d561cb8b")>
-<Assembly: AssemblyVersion("2.52.3")>
-<Assembly: AssemblyFileVersion("2.52.3")>
+<Assembly: AssemblyVersion("0.1.0")>
+<Assembly: AssemblyFileVersion("0.1.0")>
 <Assembly: AssemblyDescription("Video encoding app")>